refactor: move to ads subpackage and rename fingerprint
