+++ patches/src/main/kotlin/app/revanced/patches/angulus/ads/Fingerprints.kt
@@ -1,4 +1,4 @@
-package app.revanced.patches.angulus
+package app.revanced.patches.angulus.ads
 
 import app.revanced.patcher.fingerprint
 import com.android.tools.smali.dexlib2.AccessFlags
@@ -11,7 +11,7 @@ import com.android.tools.smali.dexlib2.AccessFlags
 
 // This fingerprint targets a method that returns the daily measurement count.
 // This method is used to determine if the user has reached the daily limit of measurements.
-internal val angulusAdsFingerprint = fingerprint {
+internal val getDailyMeasurementCountFingerprint = fingerprint {
     accessFlags(AccessFlags.PRIVATE)
     returns("I")
     strings("dailyMeasurementCount")
